feat: polishing motorcycle list page
