--- lib/features/catalog/presentation/widgets/trivia_controls.dart
@@ -1,66 +1,0 @@
-import 'package:flutter/material.dart';
-
-
-class TriviaControls extends StatefulWidget {
-  const TriviaControls({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  _TriviaControlsState createState() => _TriviaControlsState();
-}
-
-class _TriviaControlsState extends State<TriviaControls> {
-  final controller = TextEditingController();
-  late String inputStr;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: <Widget>[
-        TextField(
-          controller: controller,
-          keyboardType: TextInputType.number,
-          decoration: const InputDecoration(
-            border: OutlineInputBorder(),
-            hintText: 'Input a number',
-          ),
-          onChanged: (value) {
-            inputStr = value;
-          },
-          onSubmitted: (_) {
-            dispatchConcrete();
-          },
-        ),
-        const SizedBox(height: 10),
-        Row(
-          children: <Widget>[
-            Expanded(
-              child: TextButton(
-                // color: Theme.of(context).colorScheme.secondary,
-                // textTheme: ButtonTextTheme.primary,
-                onPressed: dispatchConcrete,
-                child: const Text('Search'),
-              ),
-            ),
-            const SizedBox(width: 10),
-            Expanded(
-              child: ElevatedButton(
-                onPressed: dispatchRandom,
-                child: const Text('Get random trivia'),
-              ),
-            ),
-          ],
-        )
-      ],
-    );
-  }
-
-  void dispatchConcrete() {
-    controller.clear();
-  }
-
-  void dispatchRandom() {
-    controller.clear();
-  }
-}
