+++ lib/features/catalog/presentation/widgets/widgets.dart
@@ -1,4 +1,3 @@
+export 'grid_builder_widget.dart';
 export 'loading_widget.dart';
 export 'message_display.dart';
-export 'trivia_display.dart';
-export 'trivia_controls.dart';
