+++ lib/injection_container.dart
@@ -26,8 +26,7 @@ Future<void> init() async {
   // getIt.registerLazySingleton(() => GetMotorcycleById(getIt()));
   getIt.registerLazySingleton(() => GetCategories(getIt()));
   getIt.registerLazySingleton(() => GetMakers(getIt()));
-  getIt.registerLazySingleton<GetMotorcycleByMaker>(
-      () => GetMotorcycleByMaker(getIt()));
+  getIt.registerLazySingleton(() => GetMotorcycleByMaker(getIt()));
 
   // Repository
   getIt.registerLazySingleton<CatalogRepository>(
